iris_hub_api/requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-PyMongo
flask-cors
APScheduler
PyGithub
humanize
gitpython
uwsgi
python-dateutil
gunicorn